fix(legal): the Impressum check fired on every filled deployment
Step 4 of docs/deploy/legal-pages.md told an operator to grep the built
bundle for "BITTE AUSFÜLLEN" and treat a hit as an unfilled Impressum.
That string is PLACEHOLDER_PREFIX, the constant isPlaceholder() compares
against, so it is compiled into EVERY bundle whether or not a field uses
it. Verified against the live glyphoxa.com bundle (v0.5.1, correctly
filled): exactly one occurrence, the constant declaration itself. A
pre-launch safety check that fires on 100% of deployments is worse than
none — it teaches operators to ignore it.

Move the check into code. `npm run check:operator` runs the very
operatorTodos() the pages call for their "Hinweis an den Betreiber"
banner, so a green run and a clean page are one fact rather than two
that can drift. It also sees what no grep over a minified bundle can: a
required field left EMPTY, and an optional field left on template text
(which renders red inline without raising the banner). Its assertions
stay fork-safe — a deployment that filled in its OWN identity passes
unchanged — so the shipped identity stays pinned in legal.test.tsx.

The bundle greps remain, for the artifact rather than the source, and
now match the colon form "BITTE AUSFÜLLEN:" that only a real placeholder
value carries. Two further bugs in them: the local one scanned all of
dist/assets/, where emptyOutDir:false leaves earlier builds that would
answer for the one that ships (it now takes the bundle index.html
references, mirroring the remote form), and neither could catch a bundle
built BEFORE operator.ts was edited — still serving the previous
operator's Impressum — so the remote block gained a positive legalName
probe.

Also corrects the section intro, which still claimed operator.ts "ships
as clearly-marked placeholders": it has shipped the filled canonical
identity since #528, and a fork must replace it rather than fill it in.

Relates to #574, #575.
